feat(*): 代码写完 没测试

+++ src/main/java/com/fullcolor/tcp/msg/resp/LoginRespData.java
@@ -31,7 +31,7 @@ public class LoginRespData implements OutboundMsg {
     }
 
     @Override
-    public short getMessageSequence() {
+    public int getMessageSequence() {
         return 1;
     }
 }
